created auth middleware to use in private routes; added auth middleware in resources that create, change or delete posts and comments

+++ backend/src/routes/comment_routes.ts
@@ -4,6 +4,7 @@ import Comment from "../core/contracts/comment";
 import CommentableBLogRepository from "../core/contracts/commentable_blog_repository";
 import CommentImpl from "../core/entities/comment_impl";
 import AppError from "../core/errors/app_error";
+import authMiddleware from "../core/middlewares/auth_middleware";
 
 
 export const commentRoutes: Router = Router();
@@ -40,13 +41,14 @@ commentRoutes.get('/allComments', async (request: Request, response: Response, n
     }
 });
 
-commentRoutes.post('/posts/:postId/comments', async (request: Request, response: Response, next: NextFunction) => {
+commentRoutes.post('/posts/:postId/comments', authMiddleware, async (request: Request, response: Response, next: NextFunction) => {
     try {
         const {postId} = request.params
-        const {text, userId} = request.body;
+        const {text} = request.body;
+        const {userId} = request.headers;
         if(!text) throw new AppError('O comentário precisa de um texto', 400)
 
-        const comment: Comment = await repository.createComment(postId, new CommentImpl(text, userId));
+        const comment: Comment = await repository.createComment(postId, new CommentImpl(text, userId as string));
 
         return response.status(201).json(comment)
     } catch (error) {
@@ -54,9 +56,16 @@ commentRoutes.post('/posts/:postId/comments', async (request: Request, response:
     }
 })
 
-commentRoutes.delete('/posts/:postId/comments/:commentId', async (request: Request, response: Response, next: NextFunction) => {
+commentRoutes.delete('/posts/:postId/comments/:commentId', authMiddleware, async (request: Request, response: Response, next: NextFunction) => {
     try {
         const {postId, commentId} = request.params
+        const {userId} = request.headers;
+        const comment: Comment = await repository.retrieveCommnent(postId, commentId);
+
+        if(comment.SUserId !== userId) {
+            throw new AppError('Você não pode apagar posts que não são seus!', 401);
+        }
+
         await repository.deleteComment(postId, commentId)
         return response.status(204).send();
     } catch (error) {
@@ -64,13 +73,20 @@ commentRoutes.delete('/posts/:postId/comments/:commentId', async (request: Reque
     }
 })
 
-commentRoutes.put('/posts/:postId/comments/:commentId', async (request: Request, response: Response, next: NextFunction) => {
+commentRoutes.put('/posts/:postId/comments/:commentId', authMiddleware, async (request: Request, response: Response, next: NextFunction) => {
     try {
-        const {postId, commentId} = request.params
-        const {text} = request.body
+        const {postId, commentId} = request.params;
+        const {text} = request.body;
+        const {userId} = request.headers;
+
         if(!text) throw new AppError('O comentário precisa de um texto', 400)
 
         const comment = await repository.retrieveCommnent(postId, commentId);
+        
+        if(comment.SUserId !== userId) {
+            throw new AppError('Você não pode apagar posts que não são seus!', 401);
+        }
+
         comment.text = text;
         await repository.updateComment(postId, comment);
         return response.status(200).send();
